complete UI and add two components NewPOst and NavBar
