+++ public/index.html
@@ -9,6 +9,11 @@
       content="Web site created using create-react-app"
     />
     <title>Face Detection and Tagging App by Shahid</title>
+    <style>
+      *{
+        margin: 0;
+      }
+    </style>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
