+++ src/App.css
@@ -6,3 +6,63 @@ canvas{
     /* background-color: red; */
     position: absolute;
 }
+
+.navbar {
+    height: 50px;
+    background-color: #4064ac;
+    color: white;
+    font-weight: bold;
+    font-size: 20px;
+    display: flex;
+    align-items: center;
+    padding: 0px 100px;
+}
+
+.avatar {
+    height: 50px;
+    width: 50px;
+    border-radius: 50%;
+    margin-right: 20px;
+}
+
+.newPostCard {
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.addPost {
+    display: flex;
+}
+
+.postForm {
+    display: flex;
+    flex-direction: column;
+}
+
+.addImg{
+    width: 25px;
+    cursor: pointer;
+}
+
+label {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+button {
+    border: none;
+    padding: 5px 10px;
+    border-radius: 5px;
+    background-color: teal;
+    color: white;
+    cursor: pointer;
+}
+
+.postInput {
+    height: 30px;
+    border: none;
+    margin-bottom: 10px;
+}
