+++ src/components/NavBar.jsx
@@ -1,0 +1,9 @@
+const NavBar = () => {
+    return (
+        <div className="navbar">
+            <span>LamaSocial</span>
+        </div>
+    );
+};
+
+export default NavBar;
